// src/types/metrics/internal.ts
interface MetricsApiBucket {
  count: number;
  period_start: string;
}

export interface MetricsEngagementApiResponse {
  buckets: {
    click: MetricsApiBucket[];
    click_tracking_delivered: MetricsApiBucket[];
    open: MetricsApiBucket[];
    open_tracking_delivered: MetricsApiBucket[];
  };
  click: number;
  click_tracking_delivered: number;
  end_time: string;
  open: number;
  open_tracking_delivered: number;
  start_time: string;
}

export interface MetricsPerformanceApiResponse {
  bounced: number;
  buckets: {
    bounced: MetricsApiBucket[];
    delivered: MetricsApiBucket[];
    processed: MetricsApiBucket[];
  };
  delivered: number;
  end_time: string;
  processed: number;
  start_time: string;
}

export interface MetricsRecipientBehaviourApiResponse {
  buckets: {
    unsubscribe_delivered: MetricsApiBucket[];
    unsubscribed: MetricsApiBucket[];
  };
  end_time: string;
  start_time: string;
  unsubscribe_delivered: number;
  unsubscribed: number;
}

export interface MetricsSenderApiResponse {
  end_time?: string;
  limit: number;
  offset: number;
  senders: {
    bounced: number;
    delivered: number;
    dropped: number;
    name: string;
    processed: number;
  }[];
  start_time?: string;
  total: number;
}

export interface MetricsVolumeApiResponse {
  buckets: {
    delivered: MetricsApiBucket[];
    dropped: MetricsApiBucket[];
    processed: MetricsApiBucket[];
  };
  delivered: number;
  dropped: number;
  end_time: string;
  processed: number;
  start_time: string;
}

export interface MetricsUsageApiResponse {
  period_end_date: string;
  period_start_date: string;
  total_usage: number;
}

export interface MetricsSendersApiResponse {
  end_time: string;
  limit: number;
  offset: number;
  senders: {
    bounced: number;
    delivered: number;
    dropped: number;
    name: string;
    processed: number;
  }[];
  start_time: string;
  total: number;
}